+++ src/node/index.ts
@@ -1,4 +1,5 @@
 export * from '../common';
 export * from './EnvVars';
 export * from './getDirname';
+export * from './getFileFingerprint';
 export type * from './TypedEventEmitter';